+++ lwsapp-cxf-client/src/main/java/com/learnwebservices/hello/HelloMain.java
@@ -12,8 +12,10 @@ import java.net.URL;
 public class HelloMain {
 
     public static void main(String[] args) throws MalformedURLException {
-        URL url = new URL("http://www.learnwebservices.com/services/hello?wsdl");
-        HelloEndpointService service = new HelloEndpointService(url);
+//        URL url = new URL("http://www.learnwebservices.com/services/hello?wsdl");
+//        HelloEndpointService service = new HelloEndpointService(url);
+
+        HelloEndpointService service = new HelloEndpointService();
         HelloEndpoint port = service.getHelloEndpointPort();
         HelloRequest request = new HelloRequest();
         request.setName("John Doe");