+++ lwsapp-js-client/app.js
@@ -1,8 +1,8 @@
 const soap = require("soap");
-const url = "http://www.learnwebservices.com/services/hello?wsdl";
-const args = {HelloRequest: {Name: "John Doe"}};
+const url = "https://apps.learnwebservices.com/services/hello?wsdl";
+const args = {Name: "John Doe"};
 soap.createClient(url, function(err, client) {
     client.SayHello(args, function(err, result) {
-        console.log(result.HelloResponse.Message);
+        console.log(result.Message);
     });
 });