+++ lwsapp-php-client/hello-client.php
@@ -1,13 +1,11 @@
 <?php
 
-$client = new SoapClient('http://www.learnwebservices.com/services/hello?wsdl');
+$client = new SoapClient('https://apps.learnwebservices.com/services/hello?wsdl');
 echo $client
 	->SayHello(
-		[
-			'HelloRequest' => 
-				['Name' => 'John Doe']
+		[			
+			'Name' => 'John Doe'
 		]
 	)
-	->HelloResponse
-		->Message
+	->Message
 ;