+++ lwsapp-python-client/hello-client.py
@@ -1,7 +1,7 @@
 import zeep
 
-wsdl = 'http://www.learnwebservices.com/services/hello?wsdl'
+wsdl = 'https://apps.learnwebservices.com/services/hello?wsdl'
 client = zeep.Client(wsdl=wsdl)
-request = {'Name': 'John Doe'}
+request = 'John Doe'
 response = client.service.SayHello(request)
 print(response)